+++ src/components/pages/movie-page-in-list.tsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import {Link, useNavigate, useParams} from 'react-router-dom';
-import {AppRoute, getReviewRoute} from '../app';
-import {FilmCards, getSimilarMovies} from '../film-card';
+import {AppRoute} from '../app';
+import {FilmCards} from '../film-card';
 import {Footer} from '../footer';
 import {Films} from '../../mocks/films';
 import {Details} from '../../mocks/details';
 import {Overviews} from '../../mocks/overview';
+import {getReviewRoute} from '../functions/get-review-route';
+import {getSimilarMovies} from '../functions/get-similar-movies';
 
 export function MoviePageInList() {
   const { id } = useParams();
