+++ src/index.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import {Provider} from 'react-redux';
 import {App} from './components/app';
+import store from './store/index';
 import {Films} from './mocks/films';
 import {Details} from './mocks/details';
 
@@ -10,10 +12,12 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App mainParams={{
-      film: Films[0],
-      detail: Details[0]
-    }}
-    />
+    <Provider store={store}>
+      <App mainParams={{
+        film: Films[0],
+        detail: Details[0]
+      }}
+      />
+    </Provider>
   </React.StrictMode>
 );
