+++ src/store/index.ts
@@ -1,0 +1,10 @@
+import {configureStore} from '@reduxjs/toolkit';
+import {appReducer} from './reducer';
+
+const store = configureStore({
+  reducer: {
+    app: appReducer
+  },
+});
+
+export default store;
